Fixes eslint error
