+++ src/components/NotificationSnackBar.tsx
@@ -1,40 +1,6 @@
-import {useState} from "react"
 import Snackbar from "@mui/material/Snackbar";
-import Alert, { AlertColor } from "@mui/material/Alert";
-
-type Notification = {
-  message?: string,
-  severity?: AlertColor,
-};
-
-type NotificationState = {
-  notification?: Notification,
-  isOpen: boolean,
-}
-
-export function useNotifications() {
-  const [current, setCurrentNotification] = useState<NotificationState>({
-    isOpen: false,
-  });
-
-  return {
-    show: (notification: Notification) => {
-      setCurrentNotification({
-        notification,
-        isOpen: true,
-      });
-    },
-
-    close: () => {
-      setCurrentNotification((value) => ({
-        notification: value.notification,
-        isOpen: false,
-      }));
-    },
-
-    current: current,
-  };
-}
+import Alert from "@mui/material/Alert";
+import { NotificationState } from "@/hooks/notifications"
 
 export type NotificationSnackBarProps = {
   currentNotification: NotificationState,
